requirements.txt
flask
gunicorn
line-bot-sdk==1.3.0
requests
opencc-python-reimplemented